move some routines funcs to controller
