--- openbb_terminal/account/account_model.py
@@ -1,166 +1,0 @@
-import os
-from pathlib import Path
-from typing import Dict, List, Optional, Tuple, Union
-
-import numpy as np
-import pandas as pd
-
-from openbb_terminal.core.session.current_user import get_current_user
-from openbb_terminal.rich_config import console
-
-__login_called = False
-
-
-def get_login_called():
-    """Get the login/logout called flag.
-
-    Returns
-    -------
-    bool
-        The login/logout called flag.
-    """
-    return __login_called
-
-
-def set_login_called(value: bool):
-    """Set the login/logout called flag.
-
-    Parameters
-    ----------
-    value : bool
-        The login/logout called flag.
-    """
-    global __login_called  # pylint: disable=global-statement
-    __login_called = value
-
-
-def get_default_routines_info(routines: List[Dict[str, str]]) -> pd.DataFrame:
-    """Get the routines list.
-
-    Parameters
-    ----------
-    response : requests.Response
-        The response.
-
-    Returns
-    -------
-    Tuple[pd.DataFrame, int, int]
-        The routines list, the current page and the total number of pages.
-    """
-    df = pd.DataFrame()
-    if routines:
-        df = pd.DataFrame(routines)
-        df = df[["name", "description", "version", "date_updated"]]
-        df.index = np.arange(1, len(df) + 1)
-    return df
-
-
-def get_personal_routines_info(response) -> Tuple[pd.DataFrame, int, int]:
-    """Get the routines list.
-
-    Parameters
-    ----------
-    response : requests.Response
-        The response.
-
-    Returns
-    -------
-    Tuple[pd.DataFrame, int, int]
-        The routines list, the current page and the total number of pages.
-    """
-    df = pd.DataFrame()
-    page = 1
-    pages = 1
-    if response and response.status_code == 200:
-        data = response.json()
-        page = data.get("page", 1)
-        pages = data.get("pages", 1)
-        items = data.get("items", [])
-        if items:
-            df = pd.DataFrame(items)
-            df.index = np.arange(1, len(df) + 1)
-
-    return df, page, pages
-
-
-def read_routine(file_name: str, folder: Optional[Path] = None) -> Optional[str]:
-    """Read the routine.
-
-    Parameters
-    ----------
-    file_name : str
-        The routine.
-    folder : Optional[Path]
-        The routines folder.
-
-    Returns
-    -------
-    file_name : str
-        The routine.
-    folder : Optional[Path]
-        The routines folder.
-    """
-
-    current_user = get_current_user()
-    if folder is None:
-        folder = current_user.preferences.USER_ROUTINES_DIRECTORY
-
-    try:
-        user_folder = folder / "hub"
-        file_path = (
-            user_folder / file_name
-            if os.path.exists(user_folder / file_name)
-            else folder / file_name
-        )
-
-        with open(file_path) as f:
-            routine = "".join(f.readlines())
-        return routine
-    except Exception:
-        console.print("[red]Failed to find routine.[/red]")
-        return None
-
-
-def save_routine(
-    file_name: str,
-    routine: str,
-    folder: Optional[Path] = None,
-    force: bool = False,
-) -> Union[Optional[Path], str]:
-    """Save the routine.
-
-    Parameters
-    ----------
-    file_name : str
-        The routine.
-    routine : str
-        The routine.
-    folder : Path
-        The routines folder.
-    force : bool
-        Force the save.
-
-    Returns
-    -------
-    Optional[Path, str]
-        The path to the routine or None.
-    """
-
-    current_user = get_current_user()
-    if folder is None:
-        folder = current_user.preferences.USER_ROUTINES_DIRECTORY
-
-    try:
-        user_folder = folder / "hub"
-        if not os.path.exists(user_folder):
-            os.makedirs(user_folder)
-
-        file_path = user_folder / file_name
-        if os.path.exists(file_path) and not force:
-            return "File already exists"
-        with open(file_path, "w") as f:
-            f.write(routine)
-        return user_folder / file_name
-    except Exception:
-        console.print("[red]\nFailed to save routine.[/red]")
-        return None
